refactor(web): decompose the engine-adapter god-files into <200-line modules (migration wave 3)
`client.ts` (2157 lines, one ~100-method `HoustonClient`) and `control-plane.ts`
(1505 lines, ~85 functions) both broke the repo's hard 200-line/file rule. They
are now thin roots over cohesive modules; the public surface is byte-identical,
so callers and the `vi.mock("…/control-plane")` test suite are untouched — a pure
structural, behavior-preserving refactor.

- 200-line rule fixed: client.ts 2157→85, control-plane.ts 1505→57; every new
  file under `client/` and `cp/` is ≤200 lines (extracted, never compressed).
- AdapterContext shared-state seam: `HoustonClient` is composed from cluster
  mixins (`client/*-mixin.ts`) over ONE `AdapterContext` (`client/context.ts`,
  held by `client/base.ts` as `protected this.ctx`). The context is the single
  source of truth for engine/sdk/authFetch/login state and the engine-routing
  helpers. `cp` is a getter over a privately-held ControlPlaneConfig; setActiveOrg
  mutates that object in place, so authFetch, the per-agent runtime clients, and
  the SDK all reroute with no rebuild and no per-mixin copy can drift. Mixins call
  control-plane only through the `../control-plane` barrel (the mock seam) and
  reach state only via `this.ctx`. control-plane.ts is now a barrel over `cp/*`
  (fetch = shared transport; runtime-clients, agent-color, agents, credentials,
  board, skills, marketplace, files-context, events, orgs, spaces-billing,
  agent-teams, integrations).
- Proxy silent-failure fix: removed the catch-all Proxy whose `get` returned
  `async () => []` for any undefined method (it masked typos AND removed legacy
  features as "empty result"). Legacy desktop/Rust-engine methods that don't
  exist on the host engine (worktrees, shell, phone tunnel/pairing, the
  Claude/Composio/Gemini CLIs) now throw an explicit error
  (client/legacy-unsupported-mixin.ts); a genuinely undefined method is now a real
  TypeError, never a silent []. New test: tests/legacy-unsupported.test.ts.
- Behavior-preserving proof: full houston-web suite green (32 files, 147 tests),
  typecheck clean, biome clean, check:boundaries clean.
