+++ packages/web/src/engine-adapter/client.ts
@@ -1,2157 +1,85 @@
-import { agentFileEventType, migrateProviderModel } from "@houston/domain";
-import type { ProviderCatalog } from "@houston/protocol";
-import {
-  type CustomEndpoint,
-  EngineError,
-  HoustonEngineClient,
-  type ProviderId,
-} from "@houston/runtime-client";
-import type { BoardStatus, HoustonSdk } from "@houston/sdk";
-import { historyToFeed as sdkHistoryToFeed } from "@houston/sdk";
-import type {
-  Activity,
-  ActivityUpdate,
-  Agent,
-  Capabilities,
-  ChatHistoryEntry,
-  CommunitySkill,
-  CommunitySkillPreview,
-  ConversationEntry,
-  CreateAgent,
-  CreateAgentResult,
-  CreateSkillRequest,
-  GenerateInstructionsResult,
-  InstallCommunityRequest,
-  InstalledConfig,
-  InstallFromGithub,
-  InstallFromRepoRequest,
-  NewActivity,
-  NewRoutine,
-  PendingInteraction,
-  PortableAnonymizeRequest,
-  PortableAnonymizeResponse,
-  PortableExportRequest,
-  PortableInstalledAgent,
-  PortableInstallRequest,
-  PortableInventoryPreview,
-  PortableScanResponse,
-  PortableUploadPreviewResponse,
-  ProjectConfig,
-  ProjectFile,
-  ProviderStatus,
-  RepoSkill,
-  Routine,
-  RoutineRun,
-  RoutineUpdate,
-  SaveSkillRequest,
-  SessionStartRequest,
-  SessionStartResponse,
-  SidebarLayout,
-  SkillDetail,
-  TunnelCredentials,
-  UpdateAgent,
-  Workspace,
-  WorkspaceContext,
-} from "../../../../ui/engine-client/src/types";
-import * as activities from "./activities";
-import {
-  readAgentFile as readAgentFileStore,
-  writeAgentFile as writeAgentFileStore,
-} from "./agent-files";
-import * as agents from "./agents";
-import { bus, emitEvent, emitLocalEcho } from "./bus";
-import type { ControlPlaneConfig } from "./control-plane";
-import * as controlPlane from "./control-plane";
-import {
-  type CachedFrame,
-  conversationCacheScope,
-  deleteCachedConversation,
-  readCachedConversation,
-  setConversationCacheIdentity,
-  writeCachedConversation,
-} from "./conversation-cache";
-import * as portable from "./portable";
-import { createEngineSdk } from "./sdk-client";
-import {
-  flushQueuedSends,
-  maybeQueueSend,
-  removeQueuedSend,
-} from "./send-queue";
-import {
-  configWriteToSettings,
-  credentialSiblings,
-  DEFAULT_AGENT_ID,
-  DEFAULT_AGENT_PATH,
-  DEFAULT_WORKSPACE_ID,
-  syntheticWorkspace,
-  toNewProvider,
-  toOldProvider,
-  wireTurnPin,
-} from "./synthetic";
-import { historyToFeed, isConversationNotFound } from "./translate";
-import {
-  observeConversation,
-  seedConversationVm,
-  streamTurn,
-} from "./turn-stream";
-
-export interface HoustonClientOptions {
-  baseUrl: string;
-  token: string;
-  /** When true, route agents + chat through the Houston control plane (cloud). */
-  controlPlane?: boolean;
-}
-
-export class HoustonEngineError extends Error {
-  constructor(
-    public status: number,
-    public body: unknown,
-  ) {
-    // Carry the host's own explanation into the message: the v3 host answers
-    // errors as `{error: "reason"}` (some routes as `{error: {message}}`).
-    // Dropping it here would reduce every failure to "engine error <status>"
-    // in the toast/log/Sentry report — the status code without the reason.
-    const detail = (body as { error?: unknown } | null)?.error;
-    const reason =
-      typeof detail === "string"
-        ? detail
-        : typeof (detail as { message?: unknown } | null)?.message === "string"
-          ? (detail as { message: string }).message
-          : undefined;
-    super(
-      reason ? `${reason} (engine error ${status})` : `engine error ${status}`,
-    );
-    this.name = "HoustonEngineError";
-  }
-  get code(): string | undefined {
-    return (this.body as { error?: { code?: string } })?.error?.code;
-  }
-  get kind(): string | undefined {
-    return (this.body as { error?: { kind?: string } })?.error?.kind;
-  }
-}
-export function isHoustonEngineError(e: unknown): e is HoustonEngineError {
-  return e instanceof HoustonEngineError;
-}
-
-/**
- * `activeLogins` key segment for a login started before any agent existed
- * (first-run: it runs in the host's hidden setup runtime, not an agent's).
- */
-const SETUP_LOGIN_KEY = "__setup__";
-
-/** The two workspace-root context files backing Settings on local/self-host
- *  (HOU-711). In cloud the same two blobs live in Supabase, not on the volume. */
-const WORKSPACE_MD = "WORKSPACE.md";
-const USER_MD = "USER.md";
-
-/**
- * localStorage key persisting the selected agent (`setPreference("last_agent_id")`).
- * `providerEngine()` routes provider connects by it, so it must never name an
- * agent the host doesn't have — see `dropLastAgentPref`.
- */
-const LAST_AGENT_PREF = "houston.pref.last_agent_id";
-
-/**
- * Treat a 404 on login/cancel as benign: it means no login was pending (or an
- * older host lacks the cancel route), so cancel's postcondition — the login
- * slot is free — already holds. The reconnect card's every press goes
- * cancel → launch, so propagating this 404 aborted the chain and the login
- * never launched (HOU-676). Every other failure still propagates.
- */
-function benignCancelMiss(e: unknown): void {
-  if (e instanceof EngineError && e.status === 404) return;
-  throw e;
-}
-
-/**
- * Preference keys that are ACCOUNT state, not device state. The engine acts on
- * them — the host scheduler fires routines in `timezone` (hosted mode stamps it
- * onto each agent's environment), `locale` backs the workspace wire shape, and
- * the legal/migration flags must survive a reinstall — so they live behind the
- * host's `/v1/preferences/:key`, never in this browser's localStorage. A
- * device-local copy is invisible to the scheduler: routines then fire in the
- * host's zone while the UI renders the browser's, an hours-off "next run"
- * (HOU-732). Everything else (theme, last_agent_id, recent models, …) is
- * per-device UI state and stays local.
- */
-const ACCOUNT_PREF_KEYS = new Set([
-  "timezone",
-  "locale",
-  "legal_acceptance",
-  "migration_reconnect_dismissed",
-]);
-
-function readLocalPref(key: string): string | null {
-  try {
-    return localStorage.getItem(`houston.pref.${key}`);
-  } catch {
-    return null; /* storage disabled */
-  }
-}
-
-function removeLocalPref(key: string): void {
-  try {
-    localStorage.removeItem(`houston.pref.${key}`);
-  } catch {
-    /* storage disabled */
-  }
-}
-
-const SIDEBAR_LAYOUT_PREF = "houston.sidebar-layout";
-const EMPTY_SIDEBAR_LAYOUT: SidebarLayout = {
-  groups: [],
-  ungroupedOrder: [],
-};
-
-/** Pure-local (no control plane) sidebar-layout persistence, mirroring how this
- *  adapter keeps other preferences in `localStorage`. */
-function readLocalSidebarLayout(workspaceId: string): SidebarLayout {
-  try {
-    const raw = localStorage.getItem(`${SIDEBAR_LAYOUT_PREF}.${workspaceId}`);
-    return raw ? (JSON.parse(raw) as SidebarLayout) : EMPTY_SIDEBAR_LAYOUT;
-  } catch {
-    return EMPTY_SIDEBAR_LAYOUT;
-  }
-}
-function writeLocalSidebarLayout(workspaceId: string, layout: SidebarLayout) {
-  try {
-    localStorage.setItem(
-      `${SIDEBAR_LAYOUT_PREF}.${workspaceId}`,
-      JSON.stringify(layout),
-    );
-  } catch {
-    /* storage disabled */
-  }
-}
-
 /**
  * Drop-in replacement for `@houston-ai/engine-client`'s HoustonClient, backed by
  * the new TS engine. Boot/chat/auth map to the new engine; a single synthetic
  * workspace holds localStorage-backed agents, their `.houston/**` files, and
- * their boards; unsupported domains are stubbed (empty) by the Proxy fallback so
- * navigation never hits an undefined method.
+ * their boards.
+ *
+ * The ~100 methods are split into cohesive cluster **mixins** (under `client/`),
+ * composed here over ONE shared {@link HoustonClientBase} → {@link AdapterContext}
+ * (`client/context.ts`) so `cp`/`engine`/`sdk` and the active-org state have a
+ * single source of truth. The public method surface (names + signatures) is
+ * unchanged, so every caller is untouched.
+ *
+ * There is deliberately NO catch-all Proxy: the old adapter masked unknown
+ * methods with `async () => []`, a silent-failure hazard. Legacy desktop/Rust
+ * methods that don't exist on the host engine now throw explicitly
+ * ({@link LegacyUnsupportedMixin}); a genuinely undefined method throws a real
+ * TypeError instead of resolving to `[]`.
  */
-export class HoustonClient {
-  private engine: HoustonEngineClient;
-  private baseUrl: string;
-  private token: string;
-  /** Non-null in cloud mode: agents + chat go through the control plane. */
-  private cp: ControlPlaneConfig | null;
-  /**
-   * The single web-side {@link HoustonSdk} (migration wave 1). Built INERT
-   * (reactivity off — no `/v1/events` streams), sharing the same gateway auth
-   * fetch as `engine`, so its calls carry the live bearer + active-space header.
-   * Later waves delegate control-plane WRITES (agents/activities/providers/
-   * integrations/preferences) to its modules instead of re-implementing them
-   * here; reads stay on TanStack Query + the `subscribeServerEvents` bus. Held
-   * but unused this wave.
-   */
-  private readonly sdk: HoustonSdk;
-  /** In-flight cloud device-code logins, keyed `${agentId}:${providerId}` — the poll guard. */
-  private activeLogins = new Set<string>();
-  /** Per-provider auth-status pollers that translate login completion into events (local mode). */
-  private loginWatchers = new Map<string, ReturnType<typeof setInterval>>();
+export type { HoustonClientOptions } from "./client/context";
+export { HoustonEngineError, isHoustonEngineError } from "./client/errors";
 
-  constructor(opts: HoustonClientOptions) {
-    this.baseUrl = opts.baseUrl.replace(/\/+$/, "");
-    this.token = opts.token;
-    const useCp =
-      opts.controlPlane ??
-      (typeof window !== "undefined" && !!window.__HOUSTON_CP__);
-    this.cp = useCp
-      ? { baseUrl: opts.baseUrl.replace(/\/+$/, ""), token: opts.token }
-      : null;
-    // Local conversation cache (HOU-712) — cloud only, scoped per gateway +
-    // signed-in user. Reads the LIVE bearer so a token refresh keeps the same
-    // scope while a different account lands in different keys; local engines
-    // resolve null and never cache (their reads are local disk, never held).
-    const cp = this.cp;
-    setConversationCacheIdentity(() =>
-      cp
-        ? conversationCacheScope(cp.baseUrl, controlPlane.liveToken(cp.token))
-        : null,
-    );
-    // Live-token auth fetch (not a pinned `token`): hosted mode rotates the
-    // Supabase bearer mid-session, and a 401 must refresh + replay instead of
-    // surfacing (HOU-687). Outside hosted mode liveToken falls back to the
-    // captured static token, so local/static hosts are unchanged. Built ONCE
-    // and shared by `engine` and the SDK below, so `x-houston-org` has a single
-    // live source (`setActiveOrg` mutates `this.cp` in place; both re-read it).
-    const authFetch = controlPlane.gatewayAuthFetch(
-      opts.token,
-      () => this.cp?.activeOrgSlug,
-    );
-    this.engine = new HoustonEngineClient({
-      baseUrl: opts.baseUrl,
-      fetch: authFetch,
-    });
-    // The single web-side HoustonSdk (migration wave 1). INERT: reactivity is
-    // off, so constructing it opens NO stream and fires NO request — it only
-    // holds the SDK's write surface for later waves. It rides the SAME
-    // `authFetch`, so its bearer, 401-refresh, and active-space header match
-    // every other gateway call with no extra wiring.
-    this.sdk = createEngineSdk({ baseUrl: this.baseUrl, fetch: authFetch });
-    // Mark the new TS engine as the active backend so the frontend can surface
-    // new-engine-only capabilities (e.g. API-key providers like OpenCode). The
-    // Rust engine uses the real `@houston-ai/engine-client`, never this adapter,
-    // so the flag stays unset there.
-    if (typeof window !== "undefined") {
-      (
-        window as unknown as { __HOUSTON_NEW_ENGINE__?: boolean }
-      ).__HOUSTON_NEW_ENGINE__ = true;
-    }
-    // biome-ignore lint/correctness/noConstructorReturn: Proxy provides transparent fallback stubs for unsupported HoustonClient methods; callers use `new HoustonClient()` directly so a static factory would require changes across many files outside this module.
-    return new Proxy(this, {
-      get(target, prop, recv) {
-        if (prop in target || typeof prop === "symbol")
-          return Reflect.get(target, prop, recv);
-        return async () => {
-          console.warn(
-            `[engine-adapter] unsupported HoustonClient.${String(prop)}() → []`,
-          );
-          return [];
-        };
-      },
-    });
-  }
+import { ActivitiesMixin } from "./client/activities-mixin";
+import { AgentFilesMixin } from "./client/agent-files-mixin";
+import { AgentsMixin } from "./client/agents-mixin";
+import { HoustonClientBase } from "./client/base";
+import { BootMixin } from "./client/boot-mixin";
+import { ChatHistoryMixin } from "./client/chat-history-mixin";
+import { ChatSendMixin } from "./client/chat-send-mixin";
+import { ConfigPrefsMixin } from "./client/config-prefs-mixin";
+import { IntegrationsMixin } from "./client/integrations-mixin";
+import { LegacyUnsupportedMixin } from "./client/legacy-unsupported-mixin";
+import { MarketplaceMixin } from "./client/marketplace-mixin";
+import { OrgsMixin } from "./client/orgs-mixin";
+import { PortableMixin } from "./client/portable-mixin";
+import { ProjectFilesMixin } from "./client/project-files-mixin";
+import { ProviderCredentialsMixin } from "./client/provider-credentials-mixin";
+import { ProviderLoginMixin } from "./client/provider-login-mixin";
+import { ProviderStatusMixin } from "./client/provider-status-mixin";
+import { RoutinesSkillsMixin } from "./client/routines-skills-mixin";
+import { TeamsMixin } from "./client/teams-mixin";
+import { WorkspacesMixin } from "./client/workspaces-mixin";
 
-  /**
-   * Cloud mode: open the host's global reactivity stream (`/v1/events`, SSE) and
-   * fan it onto the in-process bus the UI already listens on — so an activity,
-   * routine, or skill changing server-side invalidates the right query. Tied to
-   * the EngineWebSocket connect/disconnect lifecycle (returns the unsubscribe).
-   * Standalone web mode has no host stream, so this is a no-op.
-   */
-  subscribeServerEvents(): () => void {
-    if (!this.cp) return () => {};
-    return controlPlane.subscribeEvents(this.cp, (e) => bus.emit(e));
-  }
-
-  /**
-   * Pin (or clear) the active hosted space (C8 §Workspaces bridge). Pass an org
-   * slug (the `org:`-stripped id of a `kind: "org"` workspace) to act inside
-   * that team space — every gateway call then carries `x-houston-org`, and the
-   * events stream a `?org=` query — or `null` to fall back to the personal org.
-   *
-   * Mutates the live `ControlPlaneConfig` in place: the config object is shared
-   * by every per-request `cpFetch` and by the long-lived per-agent runtime
-   * clients (whose auth-fetch re-reads it per attempt), so a switch takes
-   * effect immediately without rebuilding anything. No-op outside cloud mode
-   * (`this.cp === null`) — local/self-host hosts have no space concept.
-   *
-   * `role` is per-space, so the caller MUST re-fetch `capabilities()` after
-   * switching (C8 §capabilities); this only redirects the transport.
-   */
-  setActiveOrg(slug: string | null): void {
-    if (this.cp) this.cp.activeOrgSlug = slug;
-  }
-
-  /**
-   * The web-side {@link HoustonSdk} (migration wave 1). Exposes the SDK's write
-   * modules — `agents`, `activities`, `providers`, `integrations`,
-   * `preferences` — so later waves delegate control-plane WRITES here (matching
-   * iOS) instead of re-implementing them in this adapter. It is INERT: no
-   * `/v1/events` stream, no request until a write is dispatched; reads stay on
-   * TanStack Query + the `subscribeServerEvents` bus. Nothing consumes this yet.
-   */
-  get engineSdk(): HoustonSdk {
-    return this.sdk;
-  }
-
-  private async activeOld(): Promise<{ provider: string; model: string }> {
-    try {
-      // Cloud: providers are PER-AGENT, reached through the control-plane proxy
-      // (the per-agent runtime client carries the live token). A top-level
-      // /providers on the base client has no route and a stale token → 401.
-      const engine = this.providerEngine();
-      if (engine) {
-        // Bounded: this call sits on the BOOT path (listWorkspaces → the app's
-        // wsLoading splash), and a per-agent read against a cold/warming
-        // engine is held until the engine wakes — minutes. The value only
-        // labels the synthetic workspace, so after a short budget fall back
-        // to the defaults instead of wedging the first paint (HOU-693; same
-        // invariant as "App boot — gate chain must never hang on the engine").
-        const providers = await Promise.race([
-          engine.listProviders(),
-          new Promise<null>((r) => setTimeout(() => r(null), 4_000)),
-        ]);
-        if (providers) {
-          const active =
-            providers.find((p) => p.isActive) ??
-            providers.find((p) => p.configured);
-          if (active)
-            return {
-              provider: toOldProvider(active.id),
-              model: active.activeModel,
-            };
-        }
-      }
-    } catch {
-      /* engine unreachable / no agent selected / not authed → defaults below */
-    }
-    return { provider: "anthropic", model: "claude-sonnet-4-6" };
-  }
-
-  /** The CP agent the user has selected (persisted as last_agent_id), or null. */
-  private currentAgentId(): string | null {
-    try {
-      const id = localStorage.getItem(LAST_AGENT_PREF);
-      return id && id !== DEFAULT_AGENT_ID ? id : null;
-    } catch {
-      return null;
-    }
-  }
-  /**
-   * Forget the persisted agent selection when it names an agent the control
-   * plane no longer has (deleted last agent, wiped user data, account switch).
-   * Credentials are workspace-central (connect-once), so provider connects
-   * don't need an agent — but `providerEngine()` routes them through the
-   * remembered agent's runtime whenever this pref is set. A stale id sent
-   * first-run onboarding logins to `/agents/<dead>/auth/:pid/login` → 404
-   * "agent not found" instead of the pre-agent `/setup-runtime` surface.
-   */
-  private dropLastAgentPref(isStale: (id: string) => boolean): void {
-    try {
-      const id = localStorage.getItem(LAST_AGENT_PREF);
-      if (id && id !== DEFAULT_AGENT_ID && isStale(id))
-        localStorage.removeItem(LAST_AGENT_PREF);
-    } catch {
-      /* storage disabled — currentAgentId() reads null there anyway */
-    }
-  }
-  /** The selected agent id, or a user-facing error if none is open. */
-  private requireAgentId(): string {
-    const id = this.currentAgentId();
-    if (!id) throw new Error("Open an agent first, then connect its account.");
-    return id;
-  }
-  /** Runtime client for provider/auth calls: the selected agent's sandbox in
-   *  cloud, the single runtime locally. Before ANY agent exists (first-run
-   *  onboarding), the host's hidden SETUP runtime — provider connect must work
-   *  pre-agent, and its capture lands on the personal workspace so the agent
-   *  created next is already connected. */
-  private providerEngine(): HoustonEngineClient {
-    if (!this.cp) return this.engine;
-    const id = this.currentAgentId();
-    return id
-      ? controlPlane.runtimeClientFor(this.cp, id)
-      : controlPlane.setupRuntimeClientFor(this.cp);
-  }
-
-  // ---- meta / boot ----
-  async health() {
-    const h = await this.engine.health();
-    return { status: h.status, version: h.version, protocol: 1 } as never;
-  }
-  async version() {
-    // gatewayAuthFetch on `/v1/version` (not `this.engine.version()`): the
-    // runtime-protocol client asks `/version`, a path only the pi runtime
-    // serves — the host's and the gateway's meta surface is `/v1/version`, so
-    // the old call 404'd against every host, silently breaking the
-    // migration-reconnect probe (HOU-688). Live-bearer fetch for the same
-    // reason as capabilities() (HOU-687).
-    const res = await controlPlane.gatewayAuthFetch(
-      this.token,
-      () => this.cp?.activeOrgSlug,
-    )(`${this.baseUrl}/v1/version`);
-    if (!res.ok) {
-      const body = await res.json().catch(() => ({}));
-      throw new HoustonEngineError(res.status, body);
-    }
-    return (await res.json()) as never;
-  }
-  async capabilities(): Promise<Capabilities> {
-    // gatewayAuthFetch (not `this.engine.capabilities()`) on purpose: hosted
-    // mode rotates the Supabase bearer mid-session, so the live token is read
-    // per attempt and a 401 refreshes + replays (HOU-687).
-    const res = await controlPlane.gatewayAuthFetch(
-      this.token,
-      () => this.cp?.activeOrgSlug,
-    )(`${this.baseUrl}/v1/capabilities`);
-    if (!res.ok) {
-      const body = await res.json().catch(() => ({}));
-      throw new HoustonEngineError(res.status, body);
-    }
-    return (await res.json()) as Capabilities;
-  }
-  async listWorkspaces(): Promise<Workspace[]> {
-    const { provider, model } = await this.activeOld();
-    console.info("[engine-adapter] listWorkspaces -> 1 synthetic workspace");
-    return [syntheticWorkspace(provider, model)];
-  }
-  async listAgents(workspaceId: string): Promise<Agent[]> {
-    if (this.cp) {
-      const list = await controlPlane.listAgents(this.cp);
-      // CP agent ids are global (the list ignores workspaceId), so this list is
-      // the full truth the selection pref must exist in. Pruning here heals a
-      // stale pref at boot, before the first-run connect surface mounts.
-      this.dropLastAgentPref((id) => !list.some((a) => a.id === id));
-      return list;
-    }
-    return agents.listAgents(workspaceId);
-  }
-  async createWorkspace(req: { name?: string }): Promise<Workspace> {
-    const { provider, model } = await this.activeOld();
-    return {
-      ...syntheticWorkspace(provider, model),
-      name: req?.name || "Personal",
-    };
-  }
-  async renameWorkspace(): Promise<Workspace> {
-    const { provider, model } = await this.activeOld();
-    return syntheticWorkspace(provider, model);
-  }
-  async deleteWorkspace(): Promise<void> {}
-  async setWorkspaceLocale(
-    _id: string,
-    locale: string | null,
-  ): Promise<Workspace> {
-    const { provider, model } = await this.activeOld();
-    return { ...syntheticWorkspace(provider, model), locale };
-  }
-  async setWorkspaceProvider(): Promise<Workspace> {
-    const { provider, model } = await this.activeOld();
-    return syntheticWorkspace(provider, model);
-  }
-  // Sidebar order + grouping is per-workspace UI state, persisted to
-  // localStorage exactly like the adapter's other preferences (getPreference).
-  // Deliberately NOT host-backed: it must work regardless of the engine's
-  // version, and a stale sidecar without the route would otherwise 404 every
-  // create-group / drag write.
-  async getSidebarLayout(workspaceId: string): Promise<SidebarLayout> {
-    return readLocalSidebarLayout(workspaceId);
-  }
-  async setSidebarLayout(
-    workspaceId: string,
-    layout: SidebarLayout,
-  ): Promise<SidebarLayout> {
-    writeLocalSidebarLayout(workspaceId, layout);
-    return layout;
-  }
-  async createAgent(
-    workspaceId: string,
-    req: CreateAgent,
-  ): Promise<CreateAgentResult> {
-    if (this.cp)
-      return {
-        agent: await controlPlane.createAgent(this.cp, req.name, req.color, {
-          claudeMd: req.claudeMd,
-          seeds: req.seeds,
-        }),
-      };
-    return agents.createAgent(workspaceId, req);
-  }
-  async renameAgent(
-    workspaceId: string,
-    agentId: string,
-    newName: string,
-  ): Promise<Agent> {
-    if (this.cp) return controlPlane.renameAgent(this.cp, agentId, newName);
-    return agents.renameAgent(workspaceId, agentId, newName);
-  }
-  async updateAgent(
-    workspaceId: string,
-    agentId: string,
-    req: UpdateAgent,
-  ): Promise<Agent> {
-    if (this.cp)
-      return controlPlane.updateAgentColor(this.cp, agentId, req.color);
-    return agents.updateAgentColor(workspaceId, agentId, req.color);
-  }
-  async deleteAgent(workspaceId: string, agentId: string): Promise<void> {
-    if (this.cp) {
-      await controlPlane.deleteAgent(this.cp, agentId);
-      // The selection pref must not outlive its agent: when the deleted agent
-      // was the remembered one (and it was the last — the app re-points the
-      // pref otherwise), provider connects must fall back to the setup runtime.
-      this.dropLastAgentPref((id) => id === agentId);
-      return;
-    }
-    agents.deleteAgent(workspaceId, agentId);
-  }
-  /**
-   * Create-with-AI: one one-shot generation turn on the runtime — the selected
-   * agent's sandbox in cloud / desktop-new-engine mode (same path as
-   * summarizeActivity), the single runtime locally. The dialog's brain picker
-   * sends legacy provider/model ids; migrate them to pi ids first. No engine
-   * reachable (cloud with no agent open yet) throws — the assist step shows the
-   * real reason instead of silently producing an empty agent (HOU-660).
-   */
-  async generateAgentInstructions(
-    description: string,
-    opts: { provider?: string; model?: string; signal?: AbortSignal } = {},
-  ): Promise<GenerateInstructionsResult> {
-    const engine = this.providerEngine();
-    if (!engine)
-      throw new Error("Open an agent first, then try Create with AI again.");
-    let provider: string | undefined;
-    let model = opts.model;
-    if (opts.provider) {
-      const migrated = migrateProviderModel(opts.provider, opts.model);
-      for (const d of migrated.diagnostics)
-        console.warn(`[engine-adapter] migrated generate model: ${d.message}`);
-      provider = migrated.provider;
-      model = migrated.model;
-    }
-    const r = await engine.generateAgent(description, {
-      provider,
-      model,
-      signal: opts.signal,
-    });
-    return {
-      name: r.name,
-      instructions: r.instructions,
-      // Nothing renders these yet on the new engine; keep the wire shape so the
-      // create dialog can start consuming them without an adapter change.
-      suggestedIntegrations: r.suggestedIntegrations.map((slug) => ({
-        slug: slug.toLowerCase(),
-        displayName: slug,
-      })),
-      suggestedRoutine: r.suggestedRoutine ?? null,
-    };
-  }
-  /** The one config both deployments share: the gateway in cloud mode, the
-   *  local/self-host host otherwise — each serves `/v1/preferences/:key`. */
-  private prefConfig(): ControlPlaneConfig {
-    return this.cp ?? { baseUrl: this.baseUrl, token: this.token };
-  }
-  async getPreference(key: string): Promise<string | null> {
-    if (ACCOUNT_PREF_KEYS.has(key)) {
-      const cfg = this.prefConfig();
-      const value = await controlPlane.getPreference(cfg, key);
-      if (value !== null) return value;
-      // One-time lift of a pre-fix device-local copy: earlier builds kept
-      // account keys in localStorage only, so the host never learned them.
-      // Migrate the stored value up (and drop the local copy) rather than
-      // re-deriving it — a deliberately chosen timezone must survive.
-      const legacy = readLocalPref(key);
-      if (legacy !== null) {
-        await controlPlane.setPreference(cfg, key, legacy);
-        removeLocalPref(key);
-        return legacy;
-      }
-      return null;
-    }
-    const stored = readLocalPref(key);
-    if (stored !== null) return stored;
-    // Default to the synthetic ids so the shell auto-selects the workspace +
-    // agent on first load (otherwise no agent is current and the board is empty).
-    if (key === "last_workspace_id") return DEFAULT_WORKSPACE_ID;
-    if (key === "last_agent_id") return DEFAULT_AGENT_ID;
-    return null;
-  }
-  async setPreference(key: string, value: string): Promise<void> {
-    if (ACCOUNT_PREF_KEYS.has(key)) {
-      // Delegate the account-key WRITE to the SDK (migration wave 2a): its
-      // PreferencesClient issues the identical `PUT /v1/preferences/:key` with
-      // body `{value}` over the SAME shared gateway fetch (bearer +
-      // `x-houston-org`), and — unlike the agents/activities facades — does NOT
-      // refetch. The SDK echoes the stored value; this caller discards it, so
-      // the observable request and the `void` result are byte-identical to the
-      // old `controlPlane.setPreference`. PUTs never transient-retry in either
-      // path, so nothing is lost. The READ stays on `controlPlane.getPreference`
-      // (below): cpFetch wraps GETs in `transientRetryFetch`, which the SDK path
-      // lacks — delegating it would drop that boot-path retry resilience.
-      await this.engineSdk.preferences.set(key, value);
-      removeLocalPref(key);
-      return;
-    }
-    try {
-      localStorage.setItem(`houston.pref.${key}`, value);
-    } catch {
-      /* storage disabled */
-    }
-  }
-  async getAgentConfig(): Promise<ProjectConfig> {
-    const { provider, model } = await this.activeOld();
-    return { name: "Houston", provider, model, effort: "medium" };
-  }
-  async setAgentConfig(
-    agentPath: string,
-    config: ProjectConfig,
-  ): Promise<ProjectConfig> {
-    if (config.provider) {
-      // Migrate legacy provider+model ids to ones pi-ai accepts (the runtime's
-      // getModel throws on an unknown id → a hard-failed turn). Fail-soft: an
-      // unknown value lands on the default + records a diagnostic, never a throw.
-      const { provider, model, diagnostics } = migrateProviderModel(
-        config.provider,
-        config.model,
-      );
-      for (const d of diagnostics)
-        console.warn(`[engine-adapter] migrated agent model: ${d.message}`);
-      // Settings are PER-AGENT on the host (`/agents/:id/settings`); the host
-      // root has no `/settings` route. In cloud / desktop-new-engine mode this
-      // MUST go through the agent's runtime client (the same one activeOld()
-      // READS from) — writing via the root client silently 404s, so a model
-      // pick never persists and every turn falls back to the active provider.
-      const engine = this.cp
-        ? controlPlane.runtimeClientFor(
-            this.cp,
-            agentPath || this.requireAgentId(),
-          )
-        : this.engine;
-      await engine.setSettings({ activeProvider: provider, model });
-    }
-    // Write-through echo: the config query keys on agentPath, so the picker
-    // flips without waiting for a server round trip. See bus.emitLocalEcho.
-    emitLocalEcho("ConfigChanged", { agentPath });
-    return config;
-  }
-  // Agent-config library: templates the user installed (GitHub) that the
-  // create-agent picker merges alongside the bundled ones. Standalone web has
-  // no host to keep a library — nothing installed there is the honest answer.
-  async listInstalledConfigs(): Promise<InstalledConfig[]> {
-    if (!this.cp) return [];
-    return controlPlane.listInstalledConfigs(this.cp);
-  }
-  async installAgentFromGithub(
-    req: InstallFromGithub,
-  ): Promise<{ agentId: string }> {
-    if (!this.cp) throw new Error("Installing agents needs a cloud workspace.");
-    return controlPlane.installAgentFromGithub(this.cp, req.githubUrl);
-  }
-
-  // ---- activities (board / missions) ----
-  // Cloud: the host serves them off the agent's workspace (.houston/activity).
-  // Standalone web: localStorage-backed (no host).
-  async listActivities(agentPath: string): Promise<Activity[]> {
-    if (this.cp) return controlPlane.listActivities(this.cp, agentPath);
-    return activities.listActivities(agentPath);
-  }
-  async createActivity(
-    agentPath: string,
-    input: NewActivity,
-  ): Promise<Activity> {
-    const activity = this.cp
-      ? await controlPlane.createActivity(this.cp, agentPath, input)
-      : activities.createActivity(agentPath, input);
-    emitLocalEcho("ActivityChanged", { agentPath });
-    return activity;
-  }
-  async updateActivity(
-    agentPath: string,
-    id: string,
-    updates: ActivityUpdate,
-  ): Promise<Activity> {
-    const activity = this.cp
-      ? await controlPlane.updateActivity(this.cp, agentPath, id, updates)
-      : activities.updateActivity(agentPath, id, updates);
-    emitLocalEcho("ActivityChanged", { agentPath });
-    return activity;
-  }
-  async deleteActivity(agentPath: string, id: string): Promise<void> {
-    if (this.cp) await controlPlane.deleteActivity(this.cp, agentPath, id);
-    else activities.deleteActivity(agentPath, id);
-    // The user deleted the chat — THIS is when its locally cached transcript
-    // goes too (a server 404 alone no longer drops it, HOU-731). Missions
-    // key their conversation `activity-<id>` (see setActivityStatus).
-    if (this.cp) void deleteCachedConversation(agentPath, `activity-${id}`);
-    emitLocalEcho("ActivityChanged", { agentPath });
-  }
-
-  /**
-   * Transition a chat activity's board status, honoring cloud vs standalone mode.
-   * The board READS activities from the host in cloud mode (listActivities →
-   * control plane), so a turn's status write MUST reach the host too — a
-   * localStorage write (the standalone store) would never show up on the board
-   * and the card would hang in "running". Matches by session_key, or the
-   * `activity-<id>` convention the board uses for missions with no explicit key.
-   */
-  private async setActivityStatus(
-    agentPath: string,
-    sessionKey: string,
-    status: BoardStatus,
-    pendingInteraction: PendingInteraction | null,
-  ): Promise<void> {
-    if (!this.cp) {
-      activities.setStatusBySessionKey(
-        agentPath,
-        sessionKey,
-        status,
-        pendingInteraction,
-      );
-      // Write-through echo: this is the settle path (a turn finishing PATCHes
-      // its board status). Without it the card sticks on "running" until a
-      // server event that, in hosted mode, historically never comes.
-      emitLocalEcho("ActivityChanged", { agentPath });
-      return;
-    }
-    // This write MUST land: the turn flipped its card to "running", and a
-    // turn guarantees a terminal status on exit — a lost settle write leaves
-    // the mission visibly stuck on "running" forever. The PATCH is idempotent
-    // (fixed status + interaction), so retrying a network blip or proxy
-    // hiccup is safe. cpFetch deliberately never blind-retries writes; this
-    // caller knows its write is replay-safe.
-    const retryDelaysMs = [500, 1500, 3000];
-    for (let i = 0; ; i++) {
-      try {
-        const list = await controlPlane.listActivities(this.cp, agentPath);
-        const match = list.find(
-          (a) =>
-            a.session_key === sessionKey || `activity-${a.id}` === sessionKey,
-        );
-        if (!match) return; // transient session with no board card — nothing to update
-        // `pending_interaction: null` clears it explicitly (the host route +
-        // domain applyActivityUpdate honor null); a value records the interaction.
-        await controlPlane.updateActivity(this.cp, agentPath, match.id, {
-          status,
-          pending_interaction: pendingInteraction,
-        });
-        break;
-      } catch (err) {
-        if (i >= retryDelaysMs.length) throw err;
-        await new Promise((r) => setTimeout(r, retryDelaysMs[i]));
-      }
-    }
-    emitLocalEcho("ActivityChanged", { agentPath });
-  }
-
-  // ---- agent data files (.houston/**) ----
-  // Cloud: the host serves raw .houston docs off the agent's workspace vfs (this
-  // is what the desktop UI's board/config/learnings actually read). Standalone
-  // web: localStorage.
-  async readAgentFile(agentPath: string, relPath: string): Promise<string> {
-    if (this.cp) return controlPlane.readAgentFile(this.cp, agentPath, relPath);
-    return readAgentFileStore(agentPath, relPath);
-  }
-  async writeAgentFile(
-    agentPath: string,
-    relPath: string,
-    content: string,
-  ): Promise<void> {
-    if (this.cp) {
-      await controlPlane.writeAgentFile(this.cp, agentPath, relPath, content);
-    } else {
-      writeAgentFileStore(agentPath, relPath, content);
-    }
-    // The runtime resolves the model from its OWN settings (activeProvider +
-    // models[provider]), NOT from this .houston/config doc — which is the only
-    // thing the model picker writes. Without mirroring, picking a different model
-    // (e.g. a non-default OpenCode Go model) updates the doc but every turn keeps
-    // running the provider's default. Bridge the config write into the engine.
-    await this.syncConfigToSettings(agentPath, relPath, content);
-    // Write-through echo: files-first writes (learnings, context, config doc, …)
-    // have no dedicated event, so classify the path exactly as the host watcher
-    // does and invalidate the matching cache locally. Null (e.g. `.git/**`) skips.
-    const echoType = agentFileEventType(relPath);
-    if (echoType) emitLocalEcho(echoType, { agentPath });
-  }
-
-  /**
-   * Mirror a per-agent `config.json` write (provider + model) into the engine's
-   * settings, so a model/provider pick in the chat picker actually changes what
-   * the next turn runs. Best-effort: the doc write already succeeded, and the
-   * picker only offers connected providers, so a failure here is logged (never a
-   * silent model swap) but doesn't fail the file write.
-   */
-  private async syncConfigToSettings(
-    agentPath: string,
-    relPath: string,
-    content: string,
-  ): Promise<void> {
-    const update = configWriteToSettings(relPath, content);
-    if (!update) return;
-    try {
-      const engine = this.cp
-        ? controlPlane.runtimeClientFor(this.cp, agentPath)
-        : this.engine;
-      await engine.setSettings(update);
-    } catch (err) {
-      console.error(
-        "[engine-adapter] failed to sync the model selection to the engine:",
-        err,
-      );
-    }
-  }
-  /**
-   * Workspace + user context (HOU-711). Cloud: the two Supabase-backed blobs the
-   * gateway splices into every turn — org-wide `workspace` + the caller's `user`,
-   * never on the agent volume. Local/self-host: the two files on the agent, read
-   * through the same agent-file path the CLAUDE.md instructions use.
-   */
-  async getWorkspaceContext(agentPath: string): Promise<WorkspaceContext> {
-    if (this.cp) {
-      const [workspace, user] = await Promise.all([
-        controlPlane.getContext(this.cp, "workspace"),
-        controlPlane.getContext(this.cp, "user"),
-      ]);
-      return { workspace, user };
-    }
-    const [workspace, user] = await Promise.all([
-      this.readAgentFile(agentPath, WORKSPACE_MD),
-      this.readAgentFile(agentPath, USER_MD),
-    ]);
-    return { workspace, user };
-  }
-  /** Write ONE context slot: cloud → its gateway resource, local → its file. */
-  async setWorkspaceContextSlot(
-    agentPath: string,
-    slot: "workspace" | "user",
-    content: string,
-  ): Promise<void> {
-    if (this.cp) {
-      await controlPlane.setContext(this.cp, slot, content);
-      return;
-    }
-    await this.writeAgentFile(
-      agentPath,
-      slot === "workspace" ? WORKSPACE_MD : USER_MD,
-      content,
-    );
-  }
-  async seedAgentSchemas(): Promise<void> {}
-  async migrateAgentFiles(): Promise<void> {}
-
-  // ---- composer attachments ----
-  // Upload the dropped files into the selected agent's workspace (its durable
-  // `uploads/` folder) via the host's /agents/:id/attachments route; the
-  // runtime's clamped file tools then Read them at the relative paths returned
-  // here (the sender encodes those paths into the message), in this turn or any
-  // later conversation. Standalone web has no workspace to write into — fail loud.
-  async saveAttachments(scopeId: string, files: File[]): Promise<string[]> {
-    if (files.length === 0) return [];
-    if (!this.cp) throw new Error("Attachments need a cloud workspace.");
-    return controlPlane.saveAttachments(
-      this.cp,
-      this.requireAgentId(),
-      scopeId,
-      files,
-    );
-  }
-
-  // ---- project files (the agent's REAL workspace) ----
-  // In cloud mode the workspace is a GCS prefix served by the control plane at
-  // /agents/:id/files*. agentPath IS the agentId here (folderPath = agent.id).
-  // In synthetic/local web mode there is no real workspace, so these are inert.
-  private async cpFilesFetch(
-    agentId: string,
-    path: string,
-    init?: RequestInit,
-  ): Promise<Response> {
-    if (!this.cp)
-      throw new Error("cpFilesFetch called without a control-plane config");
-    const cp = this.cp;
-    const res = await controlPlane.gatewayAuthFetch(
-      cp.token,
-      () => cp.activeOrgSlug,
-    )(`${cp.baseUrl}/agents/${encodeURIComponent(agentId)}/${path}`, {
-      ...init,
-      headers: {
-        "Content-Type": "application/json",
-        ...init?.headers,
-      },
-    });
-    if (!res.ok)
-      throw new HoustonEngineError(
-        res.status,
-        await res.json().catch(() => ({})),
-      );
-    return res;
-  }
-  async listProjectFiles(agentPath: string): Promise<ProjectFile[]> {
-    if (!this.cp) return [];
-    return (await (
-      await this.cpFilesFetch(agentPath, "files")
-    ).json()) as ProjectFile[];
-  }
-  async readProjectFile(agentPath: string, relPath: string): Promise<string> {
-    if (!this.cp) return "";
-    const res = await this.cpFilesFetch(
-      agentPath,
-      `files/read?path=${encodeURIComponent(relPath)}`,
-    );
-    const body = (await res.json()) as { content: string; base64: boolean };
-    return body.base64 ? atob(body.content) : body.content;
-  }
-  /** Raw bytes of a workspace file (binary-safe) plus its served MIME type. */
-  async downloadProjectFile(
-    agentPath: string,
-    relPath: string,
-  ): Promise<{ blob: Blob; contentType: string }> {
-    if (!this.cp) throw new Error("downloads need a cloud workspace");
-    const res = await this.cpFilesFetch(
-      agentPath,
-      `files/download?path=${encodeURIComponent(relPath)}`,
-    );
-    return {
-      blob: await res.blob(),
-      contentType:
-        res.headers.get("content-type") ?? "application/octet-stream",
-    };
-  }
-  async deleteFile(agentPath: string, relPath: string): Promise<void> {
-    if (!this.cp) return;
-    await this.cpFilesFetch(
-      agentPath,
-      `files?path=${encodeURIComponent(relPath)}`,
-      { method: "DELETE" },
-    );
-  }
-  async renameFile(
-    agentPath: string,
-    relPath: string,
-    newName: string,
-  ): Promise<void> {
-    if (!this.cp) return;
-    await this.cpFilesFetch(agentPath, "files/rename", {
-      method: "POST",
-      body: JSON.stringify({ path: relPath, newName }),
-    });
-  }
-  async createFolder(
-    agentPath: string,
-    folderName: string,
-  ): Promise<{ created: string }> {
-    if (!this.cp) return { created: folderName };
-    return (await (
-      await this.cpFilesFetch(agentPath, "files/folder", {
-        method: "POST",
-        body: JSON.stringify({ path: folderName }),
-      })
-    ).json()) as { created: string };
-  }
-  /** Upload browser Files into the workspace (Files tab drag-drop / Browse),
-   * optionally into a subfolder. One request per file, so each request stays
-   * within the host's upload cap regardless of how many files were dropped. */
-  async uploadProjectFiles(
-    agentPath: string,
-    files: File[],
-    targetDir?: string | null,
-  ): Promise<void> {
-    if (files.length === 0) return;
-    if (!this.cp) throw new Error("Uploading files needs a connected host.");
-    for (const f of files) {
-      const contentBase64 = controlPlane.bytesToBase64(
-        new Uint8Array(await f.arrayBuffer()),
-      );
-      await this.cpFilesFetch(agentPath, "files/import", {
-        method: "POST",
-        body: JSON.stringify({
-          dir: targetDir ?? null,
-          files: [{ name: f.name, contentBase64 }],
-        }),
-      });
-    }
-  }
-  /** Move a file/folder into another folder (null = workspace root). */
-  async moveProjectFile(
-    agentPath: string,
-    relPath: string,
-    toDir: string | null,
-  ): Promise<void> {
-    if (!this.cp) throw new Error("Moving files needs a connected host.");
-    await this.cpFilesFetch(agentPath, "files/move", {
-      method: "POST",
-      body: JSON.stringify({ path: relPath, toDir }),
-    });
-  }
-  /** One zip of the workspace ("Download all") or, with `path`, of a single
-   * folder's subtree — for deployments with no local file manager to reveal
-   * in (cloud pods, web builds). */
-  async downloadProjectArchive(
-    agentPath: string,
-    path?: string,
-  ): Promise<{ blob: Blob; contentType: string }> {
-    if (!this.cp) throw new Error("Downloads need a connected host.");
-    const res = await this.cpFilesFetch(
-      agentPath,
-      `files/archive${path ? `?path=${encodeURIComponent(path)}` : ""}`,
-    );
-    return {
-      blob: await res.blob(),
-      contentType: res.headers.get("content-type") ?? "application/zip",
-    };
-  }
-
-  // ---- conversations / routines / skills (mostly empty) ----
-  async listConversations(agentPath: string): Promise<ConversationEntry[]> {
-    const agentName = agents.agentNameByPath(agentPath) ?? "Houston";
-    // The board/missions list is derived from activities; in cloud those live on
-    // the host (this.listActivities un-fakes it), not localStorage.
-    const acts = await this.listActivities(agentPath);
-    return acts.map((a) =>
-      activities.activityToConversation(a, agentPath, agentName),
-    );
-  }
-  async listAllConversations(
-    agentPaths: string[],
-  ): Promise<ConversationEntry[]> {
-    const all = await Promise.all(
-      agentPaths.map((p) => this.listConversations(p)),
-    );
-    return all.flat();
-  }
-  async listRoutines(agentPath: string) {
-    if (this.cp) return controlPlane.listRoutines(this.cp, agentPath);
-    return [];
-  }
-  async listRoutineRuns(agentPath: string) {
-    if (this.cp) return controlPlane.listRoutineRuns(this.cp, agentPath);
-    return [];
-  }
-  async listSkills(agentPath: string) {
-    if (this.cp) return controlPlane.listSkills(this.cp, agentPath);
-    return [];
-  }
-  async loadSkill(agentPath: string, name: string): Promise<SkillDetail> {
-    if (this.cp) return controlPlane.loadSkill(this.cp, agentPath, name);
-    // Standalone web has no skill backend (nothing is listed), so this is
-    // unreachable; return an empty detail rather than crash if it ever isn't.
-    return { name, title: null, description: "", version: 1, content: "" };
-  }
-
-  // Routine + skill mutations route to the host (cloud); standalone web has no
-  // routine/skill backend, so they no-op there (the UI still navigates).
-  async createRoutine(agentPath: string, input: NewRoutine): Promise<Routine> {
-    if (!this.cp) return {} as Routine;
-    const routine = await controlPlane.createRoutine(this.cp, agentPath, input);
-    emitLocalEcho("RoutinesChanged", { agentPath });
-    return routine;
-  }
-  async updateRoutine(
-    agentPath: string,
-    id: string,
-    updates: RoutineUpdate,
-  ): Promise<Routine> {
-    if (!this.cp) return {} as Routine;
-    const routine = await controlPlane.updateRoutine(
-      this.cp,
-      agentPath,
-      id,
-      updates,
-    );
-    emitLocalEcho("RoutinesChanged", { agentPath });
-    return routine;
-  }
-  async deleteRoutine(agentPath: string, id: string): Promise<void> {
-    if (!this.cp) return;
-    await controlPlane.deleteRoutine(this.cp, agentPath, id);
-    emitLocalEcho("RoutinesChanged", { agentPath });
-  }
-  /** Fire a routine on demand: the host records a routine_run and starts the turn now. */
-  async runRoutineNow(agentPath: string, routineId: string): Promise<void> {
-    if (!this.cp) throw new Error("Running a routine needs a cloud workspace.");
-    await controlPlane.runRoutineNow(this.cp, agentPath, routineId);
-    emitLocalEcho("RoutineRunsChanged", { agentPath });
-  }
-  /** Stop an in-flight routine run: the host flips the row terminal, then aborts the turn. */
-  async cancelRoutineRun(
-    agentPath: string,
-    routineId: string,
-    runId: string,
-  ): Promise<RoutineRun> {
-    if (!this.cp)
-      throw new Error("Stopping a routine run needs a cloud workspace.");
-    const run = await controlPlane.cancelRoutineRun(
-      this.cp,
-      agentPath,
-      routineId,
-      runId,
-    );
-    emitLocalEcho("RoutineRunsChanged", { agentPath });
-    return run;
-  }
-  async createSkill(req: CreateSkillRequest): Promise<void> {
-    if (!this.cp) return;
-    await controlPlane.createSkill(this.cp, req.workspacePath, {
-      name: req.name,
-      description: req.description,
-      content: req.content,
-    });
-    emitLocalEcho("SkillsChanged", { agentPath: req.workspacePath });
-  }
-  async saveSkill(name: string, req: SaveSkillRequest): Promise<void> {
-    if (!this.cp) return;
-    await controlPlane.saveSkill(this.cp, req.workspacePath, name, req.content);
-    emitLocalEcho("SkillsChanged", { agentPath: req.workspacePath });
-  }
-  async deleteSkill(workspacePath: string, name: string): Promise<void> {
-    if (!this.cp) return;
-    await controlPlane.deleteSkill(this.cp, workspacePath, name);
-    emitLocalEcho("SkillsChanged", { agentPath: workspacePath });
-  }
-  // Marketplace: skills.sh search/install + GitHub repo discovery. Standalone
-  // web has no marketplace backend — searches answer empty (the dialog shows
-  // its "unavailable" state), installs refuse loudly rather than no-op.
-  async searchCommunitySkills(
-    agentPath: string,
-    query: string,
-    signal?: AbortSignal,
-  ): Promise<CommunitySkill[]> {
-    if (!this.cp) return [];
-    return controlPlane.searchCommunitySkills(
-      this.cp,
-      agentPath,
-      query,
-      signal,
-    );
-  }
-  async previewCommunitySkill(
-    agentPath: string,
-    source: string,
-    skillId: string,
-    signal?: AbortSignal,
-  ): Promise<CommunitySkillPreview> {
-    if (!this.cp) throw new Error("Previewing skills needs a cloud workspace.");
-    return controlPlane.previewCommunitySkill(
-      this.cp,
-      agentPath,
-      source,
-      skillId,
-      signal,
-    );
-  }
-  async listSkillsFromRepo(
-    agentPath: string,
-    source: string,
-    signal?: AbortSignal,
-  ): Promise<RepoSkill[]> {
-    if (!this.cp) return [];
-    return controlPlane.listSkillsFromRepo(this.cp, agentPath, source, signal);
-  }
-  async installCommunitySkill(
-    req: InstallCommunityRequest,
-    signal?: AbortSignal,
-  ): Promise<string> {
-    if (!this.cp) throw new Error("Installing skills needs a cloud workspace.");
-    const slug = await controlPlane.installCommunitySkill(
-      this.cp,
-      req.workspacePath,
-      { source: req.source, skillId: req.skillId },
-      signal,
-    );
-    emitLocalEcho("SkillsChanged", { agentPath: req.workspacePath });
-    return slug;
-  }
-  async installSkillsFromRepo(
-    req: InstallFromRepoRequest,
-    signal?: AbortSignal,
-  ): Promise<string[]> {
-    if (!this.cp) throw new Error("Installing skills needs a cloud workspace.");
-    const installed = await controlPlane.installSkillsFromRepo(
-      this.cp,
-      req.workspacePath,
-      { source: req.source, skills: req.skills },
-      signal,
-    );
-    emitLocalEcho("SkillsChanged", { agentPath: req.workspacePath });
-    return installed;
-  }
-
-  // ---- providers (auth) ----
-  // In cloud every provider call is PER-AGENT: the user connects their OWN
-  // ChatGPT/Codex subscription to a specific agent's sandbox (its own auth.json
-  // on the PVC). Login is surfaced through the same ProviderLoginUrl/Complete bus
-  // events the desktop connect dialog already consumes, so the UI is unchanged.
-  async providerStatus(name: string): Promise<ProviderStatus> {
-    return (await this.providerStatuses([name]))[0];
-  }
-  /**
-   * Batched provider status: ONE `listProviders()` round-trip, then derive every
-   * requested provider's status from it.
-   *
-   * `listProviders` already returns EVERY provider (with its configured flag and
-   * dynamic model id — the OpenAI-compatible provider's model is absent from the
-   * static catalog, so this is the picker's only source). The old per-card
-   * `providerStatus` fetched that whole list and threw away all but one entry, so
-   * a settings screen with a dozen cards fired a dozen identical round-trips —
-   * each proxied to the agent's sandbox in cloud. Fetching once and mapping N
-   * cards off the result is the fix for HOU-650.
-   */
-  async providerStatuses(names: readonly string[]): Promise<ProviderStatus[]> {
-    const byId = new Map<
-      string,
-      { configured?: boolean; activeModel?: string }
-    >();
-    try {
-      const engine = this.providerEngine();
-      if (engine) {
-        for (const p of await engine.listProviders()) byId.set(p.id, p);
-      }
-    } catch {
-      /* sandbox unreachable / no agent selected → all report not-connected */
-    }
-    return names.map((name) => {
-      const pid = toNewProvider(name);
-      const p = pid ? byId.get(pid) : undefined;
-      return {
-        provider: name,
-        cliInstalled: true,
-        authState: p?.configured ? "authenticated" : "unauthenticated",
-        cliName: name,
-        installSource: "managed",
-        cliPath: null,
-        activeModel: p?.activeModel || undefined,
-      } as ProviderStatus;
-    });
-  }
-  /**
-   * pi-ai's FULL static model catalog — every provider and every runnable model
-   * with the picker/settings metadata (`GET /v1/catalog`, wire `ProviderCatalog`).
-   *
-   * Reached the SAME direct way as `capabilities()` / `version()` (a `/v1/*` GET
-   * on `baseUrl`), NOT the control plane:
-   * `/v1/catalog` is served by the PUBLIC meta surface of the LOCAL host
-   * (desktop) and the per-agent engine pod (cloud) — not the cloud control
-   * plane. The catalog is built from pi-ai's baked registry (no egress), so it's
-   * identical on desktop and inside an egress-locked pod. Live-bearer fetch so a
-   * rotated Supabase token refreshes + replays on 401 (HOU-687), as
-   * capabilities() does.
-   */
-  async getCatalog(): Promise<ProviderCatalog> {
-    const res = await controlPlane.gatewayAuthFetch(this.token)(
-      `${this.baseUrl}/v1/catalog`,
-    );
-    // No 404 tolerance: `/v1/catalog` is served by every current host AND by the
-    // e2e/standalone-web fake host (packages/fake-host serves it via the real
-    // `buildProviderCatalog`). A 404 therefore means a genuinely stale host, and
-    // silently degrading to `[]` is what shipped the packaged app with providers
-    // but zero models. Throw like every other route so the failure surfaces (the
-    // caller keeps the seed so the UI still renders — but loudly).
-    if (!res.ok) {
-      const body = await res.json().catch(() => ({}));
-      throw new HoustonEngineError(res.status, body);
-    }
-    return (await res.json()) as ProviderCatalog;
-  }
-  // `deviceAuth` is the client's "I can't catch a loopback callback" flag — the
-  // co-located desktop sends false (it CAN), remote webapps send true. It steers
-  // Codex's flow (false → browser/loopback, true → device code); Claude keys off
-  // the runtime's own headless mode regardless. Default true so a caller that
-  // omits it never asks a remote runtime for an unreachable loopback.
-  async providerLogin(
-    name: string,
-    opts?: { deviceAuth?: boolean; enterpriseDomain?: string },
-  ): Promise<void> {
-    const pid = toNewProvider(name);
-    if (!pid) throw new Error(`provider ${name} not supported`);
-    const deviceAuth = opts?.deviceAuth ?? true;
-    // GitHub Copilot: the company GitHub domain when the user chose the Company
-    // plan in the connect dialog. Undefined => Personal/github.com (and every
-    // other provider). The runtime runs the device-code flow against that GitHub.
-    const enterpriseDomain = opts?.enterpriseDomain;
-
-    if (!this.cp) {
-      // Local single runtime. Drive the legacy login dialog: `device_code`
-      // carries the code to display; `url` (loopback) and `auth_code`
-      // (headless Claude) leave `user_code` null so the dialog shows a paste
-      // field. The runtime emits no completion event, so poll and synthesize.
-      const info = await this.engine.startLogin(
-        pid,
-        deviceAuth,
-        enterpriseDomain,
-      );
-      const url = info.kind === "device_code" ? info.verificationUri : info.url;
-      const userCode = info.kind === "device_code" ? info.userCode : null;
-      // The bus event is the ONE opening path: an app-side handler (a mounted
-      // login surface, else the shell's global fallback) opens the URL via the
-      // platform opener. A direct window.open here double-opened next to those
-      // handlers — and was a silent no-op inside the desktop's WKWebView.
-      emitEvent("ProviderLoginUrl", {
-        provider: name,
-        url,
-        user_code: userCode,
-        // `auth_code` (headless Claude setup-token): `url` is only docs, so the
-        // handler must show the paste dialog, never auto-open it. `instructions`
-        // is the runtime's paste-step copy the dialog renders above the field.
-        auth_code: info.kind === "auth_code",
-        instructions: info.kind === "auth_code" ? info.instructions : undefined,
-      });
-      this.watchLoginCompletion(pid, name);
-      return;
-    }
-
-    // Control-plane path (cloud sandbox OR the desktop host sidecar). Start the
-    // login in THIS agent's runtime — or, before any agent exists (first-run
-    // onboarding connects the AI ahead of agent creation), in the host's hidden
-    // SETUP runtime — and surface it on the bus the picker/settings handler
-    // consumes. A remote runtime returns a device_code (we pass its
-    // `user_code`, which opens the code panel); a co-located desktop client gets
-    // a loopback `url` (user_code null) that the handler opens straight in the
-    // browser. `provider` MUST be the old/frontend id (the dialog's contract).
-    const agentId = this.currentAgentId();
-    const old = toOldProvider(pid);
-    const engine = agentId
-      ? controlPlane.runtimeClientFor(this.cp, agentId)
-      : controlPlane.setupRuntimeClientFor(this.cp);
-    const info = await engine.startLogin(pid, deviceAuth, enterpriseDomain);
-    if (info.kind === "device_code") {
-      emitEvent("ProviderLoginUrl", {
-        provider: old,
-        url: info.verificationUri,
-        user_code: info.userCode,
-      });
-    } else {
-      emitEvent("ProviderLoginUrl", {
-        provider: old,
-        url: info.url,
-        user_code: null,
-        // Setup-token paste flow (Claude): the url is docs-only, so the handler
-        // shows the paste dialog instead of opening it. `instructions` carries
-        // the runtime's paste-step copy; absent for the loopback `url` kind.
-        auth_code: info.kind === "auth_code",
-        instructions: info.kind === "auth_code" ? info.instructions : undefined,
-      });
-    }
-    void this.pollProviderConnect(agentId, pid, old);
-  }
-  async submitProviderLoginCode(name: string, code: string): Promise<void> {
-    const pid = toNewProvider(name);
-    if (!pid) return;
-    // Same target the login started in: the agent's runtime, or the setup
-    // runtime when first-run connected pre-agent.
-    const engine = this.cp ? this.providerEngine() : this.engine;
-    await engine.completeLogin(pid, code);
-  }
-  async cancelProviderLogin(name?: string): Promise<void> {
-    const pid = name ? toNewProvider(name) : undefined;
-    if (!name || !pid) return;
-    if (this.cp) {
-      // Key mirrors pollProviderConnect: the agent's id, or the setup-runtime
-      // sentinel when the first-run login started before any agent existed.
-      const agentId = this.currentAgentId();
-      this.activeLogins.delete(`${agentId ?? SETUP_LOGIN_KEY}:${pid}`); // stop the poll
-      // Kill the runtime-side login too, in the same runtime the login started
-      // in (the agent's sandbox, or the hidden setup runtime pre-agent) —
-      // otherwise it keeps polling the provider until timeout and a retry
-      // collides with the stale flow ("sign-in already pending", HOU-664 /
-      // the HOU-438 failure class).
-      await this.providerEngine().cancelLogin(pid).catch(benignCancelMiss);
-      return;
-    }
-    this.stopLoginWatch(name);
-    // Cancel the runtime's in-flight OAuth flow for real (frees the loopback
-    // port + login slot), not just the local watcher.
-    await this.engine.cancelLogin(pid).catch(benignCancelMiss);
-    // Benign completion: clears the dialog + spinner without an error toast,
-    // matching the old engine's cancel semantics.
-    emitEvent("ProviderLoginComplete", {
-      provider: name,
-      success: false,
-      error: null,
-    });
-  }
-
-  /**
-   * Poll auth status until the in-flight login for `pid` resolves, then emit
-   * `ProviderLoginComplete` so the legacy dialog closes and the card flips.
-   * Covers all three flows: loopback auto-catch, pasted headless code, and
-   * device-code polling. Local mode only (cloud uses pollProviderConnect).
-   */
-  private watchLoginCompletion(pid: ProviderId, name: string): void {
-    this.stopLoginWatch(name);
-    const startedAt = Date.now();
-    const finish = (success: boolean, error: string | null) => {
-      this.stopLoginWatch(name);
-      emitEvent("ProviderLoginComplete", { provider: name, success, error });
-    };
-    const timer = setInterval(() => {
-      void (async () => {
-        try {
-          const status = await this.engine.authStatus();
-          const pr = status.providers.find((p) => p.provider === pid);
-          if (pr?.configured) finish(true, null);
-          else if (pr?.login?.status === "error")
-            finish(false, pr?.login?.error ?? "Login failed");
-          else if (Date.now() - startedAt > 10 * 60 * 1000)
-            finish(false, "Login timed out");
-        } catch {
-          /* engine briefly unreachable; keep polling */
-        }
-      })();
-    }, 1500);
-    this.loginWatchers.set(name, timer);
-  }
-
-  private stopLoginWatch(name: string): void {
-    const timer = this.loginWatchers.get(name);
-    if (timer !== undefined) {
-      clearInterval(timer);
-      this.loginWatchers.delete(name);
-    }
-  }
-  async providerLogout(name: string): Promise<void> {
-    const pid = toNewProvider(name);
-    if (!pid) return;
-    // Sign-out clears every gateway the connect card represents — for OpenCode
-    // that's both Zen and Go, since one key connected both. Clearing a gateway
-    // that was never connected is a benign no-op.
-    const targets = credentialSiblings(pid);
-    if (this.cp) {
-      // Connect-once logout. Clearing only the runtime's local auth.json (what
-      // engine.logout does) is NOT enough: the credential also lives in the
-      // workspace's CENTRAL store, and the runtime re-pulls it from the host
-      // before every turn — so the next message re-hydrated the agent and the
-      // provider showed connected again. Forget the central credential FIRST so
-      // no in-flight turn can re-serve it, then clear the runtime's local copy.
-      const agentId = this.requireAgentId();
-      for (const target of targets) {
-        await controlPlane.forgetCredential(this.cp, agentId, target);
-        await controlPlane.runtimeClientFor(this.cp, agentId).logout(target);
-      }
-      return;
-    }
-    for (const target of targets) {
-      await this.engine.logout(target);
-    }
-  }
-
-  /**
-   * Push a desktop-extracted Anthropic OAuth credential (the `claude` CLI's
-   * `.credentials.json` JSON) to the given agent's pod, which stores + materializes
-   * it on the PVC. The desktop calls this for a REMOTE engine after a successful
-   * browser login — the pod can't read this machine's Keychain. Cloud-only:
-   * a co-located engine shares the credential dir with its local runtime, so it
-   * never reaches here (a call without a control plane is a programming error).
-   */
-  async pushClaudeOAuthCredential(
-    agentId: string,
-    credentialJson: string,
-  ): Promise<void> {
-    if (!this.cp) {
-      throw new Error("Pushing a Claude credential needs a cloud engine.");
-    }
-    await controlPlane.pushClaudeOAuthCredential(
-      this.cp,
-      agentId,
-      credentialJson,
-    );
-  }
-
-  /**
-   * Connect an API-key provider (OpenCode Zen / Go): the user pastes a key, no
-   * OAuth dance. Cloud stores it centrally (and pushes it into the agent runtime)
-   * via the control plane; local writes it straight to the single runtime. On
-   * success we fire `ProviderLoginComplete` so the connect dialog closes and the
-   * provider card flips to connected — the same signal the OAuth flow emits. A
-   * failure rejects so the caller surfaces the real reason (never swallowed).
-   */
-  async setProviderApiKey(name: string, apiKey: string): Promise<void> {
-    const pid = toNewProvider(name);
-    if (!pid) throw new Error(`provider ${name} not supported`);
-    // OpenCode's Zen + Go gateways share one opencode.ai key (pi reads
-    // OPENCODE_API_KEY for both), so store the pasted key under every sibling
-    // gateway — one connect lights up both. `pid` (the connected id) is the one
-    // that becomes active; the order of the writes doesn't affect that.
-    const targets = credentialSiblings(pid);
-    if (this.cp) {
-      // First-run pre-agent: store through the setup runtime instead — the key
-      // lands on the personal workspace and the agent created next reads it.
-      // No per-agent settings exist yet to flip.
-      const agentId = this.currentAgentId();
-      if (!agentId) {
-        for (const target of targets) {
-          await controlPlane.setSetupApiKey(this.cp, target, apiKey);
-        }
-        emitEvent("ProviderLoginComplete", {
-          provider: name,
-          success: true,
-          error: null,
-        });
-        return;
-      }
-      for (const target of targets) {
-        await controlPlane.setApiKey(this.cp, agentId, target, apiKey);
-      }
-      // CLAIM (don't set) the active provider: it becomes active only when the
-      // agent doesn't already resolve to one — a first connect on a fresh
-      // agent. Connecting a credential is not a model pick (HOU-695):
-      // unconditionally activating it here used to flip every open chat onto
-      // the new provider (paste an OpenCode key mid-Codex-chat → the next turn
-      // answers, bills, and quota-errors on OpenCode). Switching stays the
-      // model picker's job. Settings are PER-AGENT on the host, so this MUST
-      // go through the agent's runtime client.
-      await controlPlane
-        .runtimeClientFor(this.cp, agentId)
-        .claimActiveProvider(pid);
-    } else {
-      for (const target of targets) {
-        await this.engine.setApiKey(target, apiKey);
-      }
-      await this.engine.claimActiveProvider(pid);
-    }
-    // One completion event for the single account the user connected (never one
-    // per gateway), so the connect dialog closes and exactly one card flips.
-    emitEvent("ProviderLoginComplete", {
-      provider: name,
-      success: true,
-      error: null,
-    });
-  }
-
-  /**
-   * Connect an OpenAI-compatible (local) server: persist the base URL + model
-   * and CLAIM it as active (first connect on a fresh agent only — a connect
-   * never moves an agent that already has a provider, HOU-695), then fire
-   * `ProviderLoginComplete` like the other connect paths. LOCAL/desktop only —
-   * in cloud the host refuses (the openaiCompatible capability is off), so the
-   * error surfaces to the dialog. Settings are PER-AGENT on the host, so the
-   * claim MUST go through the agent's runtime client (mirrors setProviderApiKey).
-   */
-  async setProviderCustomEndpoint(endpoint: CustomEndpoint): Promise<void> {
-    if (this.cp) {
-      const agentId = this.requireAgentId();
-      await controlPlane.setCustomEndpoint(this.cp, agentId, endpoint);
-      await controlPlane
-        .runtimeClientFor(this.cp, agentId)
-        .claimActiveProvider("openai-compatible");
-    } else {
-      await this.engine.setCustomEndpoint(endpoint);
-      await this.engine.claimActiveProvider("openai-compatible");
-    }
-    emitEvent("ProviderLoginComplete", {
-      provider: "openai-compatible",
-      success: true,
-      error: null,
-    });
-  }
-
-  /**
-   * Mint a relay credential for the guided "connect a local model" flow: the
-   * desktop tunnels the user's local model server up to their CLOUD agent (see
-   * control-plane.getTunnelCredentials). Cloud/hosted only — locally there is no
-   * gateway to issue one (and no tunnel is needed, the runtime is co-located),
-   * so reject loudly rather than pretend.
-   */
-  async getTunnelCredentials(): Promise<TunnelCredentials> {
-    if (!this.cp)
-      throw new Error(
-        "Connecting a local model to a cloud agent needs a cloud workspace.",
-      );
-    return controlPlane.getTunnelCredentials(this.cp);
-  }
-
-  /**
-   * Poll the agent's sandbox until the device-code login lands (the runtime
-   * polls OpenAI in-process and writes auth.json to the PVC), then CLAIM the
-   * new provider as this agent's active one (first connect only — never moves
-   * an agent that already has a provider, HOU-695) and signal completion —
-   * which closes the dialog and refreshes provider status. Emits a failure on timeout (no silent
-   * stall). Cancellable via `cancelProviderLogin`. A null `agentId` is the
-   * first-run pre-agent flow: the login ran in the host's hidden SETUP runtime,
-   * so poll + capture there (no per-agent settings to flip yet — the agent
-   * created next carries its provider from creation).
-   */
-  private async pollProviderConnect(
-    agentId: string | null,
-    pid: ProviderId,
-    oldProvider: string,
-  ): Promise<void> {
-    if (!this.cp) return;
-    const key = `${agentId ?? SETUP_LOGIN_KEY}:${pid}`;
-    this.activeLogins.add(key);
-    const engine = agentId
-      ? controlPlane.runtimeClientFor(this.cp, agentId)
-      : controlPlane.setupRuntimeClientFor(this.cp);
-    const deadline = Date.now() + 5 * 60 * 1000;
-    try {
-      while (Date.now() < deadline) {
-        await new Promise((r) => setTimeout(r, 4000));
-        if (!this.activeLogins.has(key)) return; // cancelled
-        let configured = false;
-        try {
-          const s = await engine.authStatus();
-          configured =
-            s.providers.find((p) => p.provider === pid)?.configured ?? false;
-        } catch {
-          /* transient — keep polling */
-        }
-        if (configured) {
-          // CLAIM (don't set) the active provider: it becomes active only for
-          // a first connect on a fresh agent — a connect never moves an agent
-          // that already has a provider, so no open chat switches (HOU-695).
-          // Skipped pre-agent: the setup runtime has no agent settings.
-          if (agentId) {
-            try {
-              await engine.claimActiveProvider(pid);
-            } catch {
-              /* non-fatal: the user can pick the model in the chat header */
-            }
-          }
-          // Connect-once: store this credential for the WHOLE workspace, so every
-          // agent (existing + new + the one onboarding creates next) shares it.
-          try {
-            if (agentId) {
-              await controlPlane.captureCredential(this.cp, agentId, pid);
-            } else {
-              await controlPlane.captureSetupCredential(this.cp, pid);
-            }
-          } catch (e) {
-            console.error("[connect] workspace credential capture failed", e);
-          }
-          emitEvent("ProviderLoginComplete", {
-            provider: oldProvider,
-            success: true,
-            error: null,
-          });
-          return;
-        }
-      }
-      emitEvent("ProviderLoginComplete", {
-        provider: oldProvider,
-        success: false,
-        error: "Connection timed out. Please try connecting again.",
-      });
-    } finally {
-      this.activeLogins.delete(key);
-    }
-  }
-
-  // ---- sessions / chat ----
-  async startSession(
-    agentPath: string,
-    req: SessionStartRequest,
-  ): Promise<SessionStartResponse> {
-    const path = agentPath || DEFAULT_AGENT_PATH;
-    // In cloud mode, talk to this agent's sandbox via the control plane's proxy;
-    // locally, the single runtime. Either way `streamTurn` is identical.
-    const engine = this.cp
-      ? controlPlane.runtimeClientFor(this.cp, path)
-      : this.engine;
-    // Queue-while-running: a send into a conversation whose turn is still
-    // streaming is held and flushed as ONE combined send at settle (see
-    // send-queue.ts). Every send path inherits this here.
-    if (maybeQueueSend(path, req)) return { sessionKey: req.sessionKey };
-    // Fire-and-stream: events flow to the feed store over the bus/WS adapter.
-    // The board-status setter is cloud-aware (writes land where the board reads).
-    // The request's provider/model/effort (the chat's OWN pick, app dialect)
-    // ride the send as a per-turn wire pin in engine ids (wireTurnPin), so the
-    // turn runs on this conversation's provider — not the agent-wide settings
-    // some other chat or connect flow last wrote (HOU-695).
-    void streamTurn(
-      engine,
-      path,
-      req.sessionKey,
-      req.prompt,
-      (status, pendingInteraction) =>
-        this.setActivityStatus(
-          path,
-          req.sessionKey,
-          status,
-          pendingInteraction,
-        ),
-      req.provider,
-      undefined,
-      req.suppressUserBubble,
-      wireTurnPin(req),
-      req.displayText,
-    ).finally(() => {
-      // The turn settled (or failed): release anything queued behind it.
-      flushQueuedSends(path, req.sessionKey, (r) => {
-        void this.startSession(path, r);
-      });
-    });
-    return { sessionKey: req.sessionKey };
-  }
-
-  /** Drop one queued (not yet sent) message from a conversation's send queue. */
-  removeQueuedMessage(agentPath: string, sessionKey: string, id: string): void {
-    removeQueuedSend(agentPath || DEFAULT_AGENT_PATH, sessionKey, id);
-  }
-  async cancelSession(agentPath: string, sessionKey: string) {
-    const engine = this.cp
-      ? controlPlane.runtimeClientFor(this.cp, agentPath)
-      : this.engine;
-    // Abort the agent's in-flight turn. The engine reports whether a turn was
-    // ACTUALLY in flight. `false` means there was nothing to abort: the turn is
-    // orphaned — its board card is stuck "running" because the turn died without
-    // settling (an error that never reached a terminal frame, or an app restart
-    // that dropped the in-memory turn). Stop is the user's escape hatch, so in
-    // that case settle the card ourselves. A genuinely live turn (`true`) is
-    // settled by its own `streamTurn` when the abort lands, so we leave its
-    // status alone — writing it here too would race that terminal write.
-    const { cancelled } = await engine.cancel(sessionKey);
-    if (cancelled !== true) {
-      // Orphan rescue: a user Stop on a dead turn — never a pending interaction.
-      await this.setActivityStatus(agentPath, sessionKey, "needs_you", null);
-    }
-    return { cancelled: cancelled === true };
-  }
-  async startOnboarding(
-    _agentPath: string,
-    sessionKey: string,
-  ): Promise<SessionStartResponse> {
-    return { sessionKey };
-  }
-  async loadChatHistory(
-    agentPath: string,
-    sessionKey: string,
-    opts: { observe?: boolean } = {},
-  ): Promise<ChatHistoryEntry[]> {
-    // Cache-first paint (HOU-712): a cloud read is HELD by the gateway for
-    // the whole engine-pod cold start, so seed the VM from the last locally
-    // persisted transcript NOW — the chat shows its messages instantly — and
-    // let the network read below revalidate whenever it lands. The seed
-    // guards in seedConversationVm keep a live or richer VM untouched, so a
-    // stale cache can never clobber fresh state.
-    let cachedFrames: CachedFrame[] | null = null;
-    if (this.cp) {
-      cachedFrames = await readCachedConversation(agentPath, sessionKey);
-      if (cachedFrames && cachedFrames.length > 0 && opts.observe !== false) {
-        seedConversationVm(agentPath, sessionKey, cachedFrames);
-      }
-    }
-    try {
-      const engine = this.cp
-        ? controlPlane.runtimeClientFor(this.cp, agentPath)
-        : this.engine;
-      const history = await engine.getHistory(sessionKey);
-      const sdkFeed = sdkHistoryToFeed(history.messages);
-      // Refresh the local copy on EVERY successful read (bulk scans too), so
-      // the next cold open paints the freshest transcript we ever saw.
-      if (this.cp) {
-        void writeCachedConversation(agentPath, sessionKey, sdkFeed);
-      }
-      // Observer mode: a loaded chat may have a turn in flight that THIS client
-      // isn't streaming (page reloaded mid-turn, or another client sent it).
-      // Attach a passive resumable stream: if the server's `sync` reports a
-      // running turn it surfaces (spinner + partial) and renders to completion;
-      // an idle conversation closes the stream right after that `sync`. No-op
-      // when the conversation is already streamed here. `observe: false` is
-      // for BULK history reads (mission search, board scans) that load N
-      // conversations at a time and must not spawn N streams — only a real
-      // conversation open observes (the default).
-      if (opts.observe !== false) {
-        // Seed FIRST (the chat opens complete), then attach: the observer
-        // renders any in-flight turn live into the same VM. Seeding is a no-op
-        // when a live stream already owns this conversation (see
-        // seedConversationVm) — its feed IS the VM. The VM seed is the SDK's
-        // UNMAPPED fold: the VM carries engine provider ids uniformly (seeded
-        // and live alike); the app's binding hook owns the old-id remap.
-        seedConversationVm(agentPath, sessionKey, sdkFeed);
-        observeConversation(
-          engine,
-          agentPath,
-          sessionKey,
-          (status, pendingInteraction) =>
-            this.setActivityStatus(
-              agentPath,
-              sessionKey,
-              status,
-              pendingInteraction,
+/**
+ * The composed client. Mixin order is irrelevant — the clusters are
+ * method-disjoint and all state lives on the shared `ctx`, not on any mixin.
+ */
+const Composed = BootMixin(
+  WorkspacesMixin(
+    AgentsMixin(
+      ConfigPrefsMixin(
+        ActivitiesMixin(
+          AgentFilesMixin(
+            ProjectFilesMixin(
+              RoutinesSkillsMixin(
+                MarketplaceMixin(
+                  ChatSendMixin(
+                    ChatHistoryMixin(
+                      ProviderStatusMixin(
+                        ProviderLoginMixin(
+                          ProviderCredentialsMixin(
+                            IntegrationsMixin(
+                              OrgsMixin(
+                                TeamsMixin(
+                                  PortableMixin(
+                                    LegacyUnsupportedMixin(HoustonClientBase),
+                                  ),
+                                ),
+                              ),
+                            ),
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+                ),
+              ),
             ),
-          history.messages.length,
-        );
-      }
-      return historyToFeed(history.messages);
-    } catch (err) {
-      // A conversation with no persisted turns yet 404s — that IS an empty
-      // conversation (a fresh card opened before its first turn lands), not
-      // a failure. Anything else (network drop, auth, 5xx) propagates so the
-      // app's `call()` wrapper toasts it with the Report-bug affordance —
-      // returning [] would render a fake empty chat and swallow the error.
-      if (isConversationNotFound(err)) {
-        // A 404 with a locally cached transcript is NOT proof the chat never
-        // existed: an engine pod can answer 404 while its data is lost or not
-        // yet restored (volume recreation, seed self-heal window). The local
-        // copy is the user's only surviving transcript then — serve it and
-        // KEEP it (HOU-731). A truly deleted conversation drops out of the
-        // conversation list, so nothing reopens its cached ghost; the size
-        // cap prunes the orphaned entry eventually.
-        if (cachedFrames && cachedFrames.length > 0) {
-          return cachedFrames as ChatHistoryEntry[];
-        }
-        return [];
-      }
-      throw err;
-    }
-  }
-  /**
-   * Ask the engine to summarize the user's first message into a short mission
-   * title. Cloud: the per-agent runtime client (the same path other conversation
-   * calls take) runs an LLM title turn in the agent's sandbox. Local: the single
-   * runtime. A clean truncation fallback covers an empty model reply, a missing
-   * agent, or any transport failure — the title is cosmetic, never block the send.
-   */
-  async summarizeActivity(message: string, opts: { agentPath?: string } = {}) {
-    const truncated =
-      message.replace(/\s+/g, " ").trim().slice(0, 60) || "New chat";
-    try {
-      const agentId = opts.agentPath || this.currentAgentId() || undefined;
-      const engine = this.cp
-        ? agentId
-          ? controlPlane.runtimeClientFor(this.cp, agentId)
-          : null
-        : this.engine;
-      if (engine) {
-        const { title } = await engine.summarizeText(message);
-        const clean = title.trim();
-        if (clean) return { title: clean, description: "" };
-      }
-    } catch {
-      /* engine unreachable / not authed / no agent → fall back to truncation */
-    }
-    return { title: truncated, description: "" };
-  }
+          ),
+        ),
+      ),
+    ),
+  ),
+);
 
-  // ---- portable agents (share with / from a friend) — host only ----
-  // The wizards' backend. Preview/export/anonymize/install talk to the
-  // host's v3 portable routes; the uploaded archive is unpacked in the
-  // browser, parked in memory until install, and the threat scan runs on it
-  // right there — the scan is the same pure `@houston/domain` heuristic the
-  // host uses (see ./portable.ts).
-  async portablePreview(agentPath: string): Promise<PortableInventoryPreview> {
-    if (!this.cp) throw new Error("Sharing an agent needs a connected host.");
-    return portable.exportPreview(this.cp, agentPath);
-  }
-  async portablePackage(
-    agentPath: string,
-    req: PortableExportRequest,
-  ): Promise<ArrayBuffer> {
-    if (!this.cp) throw new Error("Sharing an agent needs a connected host.");
-    return portable.exportPackage(this.cp, agentPath, req);
-  }
-  async portableAnonymize(
-    agentPath: string,
-    req: PortableAnonymizeRequest,
-  ): Promise<PortableAnonymizeResponse> {
-    if (!this.cp) throw new Error("Sharing an agent needs a connected host.");
-    return portable.anonymize(this.cp, agentPath, req);
-  }
-  async importPreview(
-    bytes: ArrayBuffer | Uint8Array,
-  ): Promise<PortableUploadPreviewResponse> {
-    return portable.previewUpload(bytes);
-  }
-  async importScan(packageId: string): Promise<PortableScanResponse> {
-    return portable.scanUpload(packageId);
-  }
-  async importInstall(
-    req: PortableInstallRequest,
-  ): Promise<PortableInstalledAgent> {
-    if (!this.cp) throw new Error("Importing an agent needs a connected host.");
-    return portable.install(this.cp, req);
-  }
-
-  // ---- integrations (Composio, platform mode) — host only ----
-  async integrationStatus(): Promise<controlPlane.IntegrationProviderStatus[]> {
-    if (!this.cp) return [];
-    return controlPlane.integrationStatus(this.cp);
-  }
-  async setIntegrationSession(token: string | null): Promise<void> {
-    if (!this.cp) return;
-    return controlPlane.setIntegrationSession(this.cp, token);
-  }
-  async integrationToolkits(
-    provider: string,
-  ): Promise<controlPlane.IntegrationToolkit[]> {
-    if (!this.cp) return [];
-    return controlPlane.integrationToolkits(this.cp, provider);
-  }
-  async integrationConnections(
-    provider: string,
-  ): Promise<controlPlane.IntegrationConnection[]> {
-    if (!this.cp) return [];
-    return controlPlane.integrationConnections(this.cp, provider);
-  }
-  async connectIntegration(
-    provider: string,
-    toolkit: string,
-    agent?: string,
-  ): Promise<{ redirectUrl: string; connectionId: string }> {
-    if (!this.cp) throw new Error("Integrations require a connected host");
-    return controlPlane.connectIntegration(this.cp, provider, toolkit, agent);
-  }
-  async integrationConnection(
-    provider: string,
-    connectionId: string,
-  ): Promise<controlPlane.IntegrationConnection> {
-    if (!this.cp) throw new Error("Integrations require a connected host");
-    return controlPlane.integrationConnection(this.cp, provider, connectionId);
-  }
-  async disconnectIntegration(
-    provider: string,
-    toolkit: string,
-  ): Promise<void> {
-    if (!this.cp) return;
-    return controlPlane.disconnectIntegration(this.cp, provider, toolkit);
-  }
-  async dismissIntegrationsReconnectNotice(): Promise<void> {
-    // The notice only ever renders from a host-reported `reconnect` flag, so
-    // dismissing without a host is a real failure — surface it, don't no-op.
-    if (!this.cp) throw new Error("Integrations require a connected host");
-    return controlPlane.dismissIntegrationsReconnectNotice(this.cp);
-  }
-
-  // ---- org / roles (multiplayer) — hosted gateway only ----
-  async getOrg(): Promise<controlPlane.OrgInfo> {
-    if (!this.cp) throw new Error("multiplayer requires the hosted gateway");
-    return controlPlane.getOrg(this.cp);
-  }
-  async addOrgMember(
-    email: string,
-    role: controlPlane.OrgRole,
-  ): Promise<controlPlane.AddOrgMemberResult> {
-    if (!this.cp) throw new Error("multiplayer requires the hosted gateway");
-    return controlPlane.addOrgMember(this.cp, email, role);
-  }
-  async deleteOrgInvite(inviteId: string): Promise<void> {
-    if (!this.cp) throw new Error("multiplayer requires the hosted gateway");
-    return controlPlane.deleteOrgInvite(this.cp, inviteId);
-  }
-  async removeOrgMember(userId: string): Promise<void> {
-    if (!this.cp) throw new Error("multiplayer requires the hosted gateway");
-    return controlPlane.removeOrgMember(this.cp, userId);
-  }
-  async setOrgMemberRole(
-    userId: string,
-    role: controlPlane.OrgRole,
-  ): Promise<void> {
-    if (!this.cp) throw new Error("multiplayer requires the hosted gateway");
-    return controlPlane.setOrgMemberRole(this.cp, userId, role);
-  }
-
-  // ---- spaces / teams (C8) — hosted gateway only ----
-  // Off-cloud (`this.cp === null`) there is no space concept: `listOrgs` reports
-  // an empty result (the switcher shows only the personal workspace), while the
-  // mutating calls throw — a create/move must reach the gateway.
-  async listOrgs(): Promise<controlPlane.OrgsList> {
-    if (!this.cp) return { orgs: [], invites: [] };
-    return controlPlane.listOrgs(this.cp);
-  }
-  async createOrg(name: string): Promise<controlPlane.OrgSummary> {
-    if (!this.cp) throw new Error("Creating a team needs the hosted gateway.");
-    return controlPlane.createOrg(this.cp, name);
-  }
-  async moveAgent(
-    agentSlugOrId: string,
-    toSlug: string,
-  ): Promise<controlPlane.AgentMoveStart> {
-    if (!this.cp) throw new Error("Moving an agent needs the hosted gateway.");
-    return controlPlane.moveAgent(this.cp, agentSlugOrId, toSlug);
-  }
-  async getMoveStatus(
-    agentSlugOrId: string,
-    moveId: string,
-  ): Promise<controlPlane.AgentMoveStatus> {
-    if (!this.cp) throw new Error("Moving an agent needs the hosted gateway.");
-    return controlPlane.getMoveStatus(this.cp, agentSlugOrId, moveId);
-  }
-
-  // ---- billing (C8) — hosted gateway only ----
-  // Off-cloud (`this.cp === null`) there is no team/billing concept: the read
-  // degrades to null (the billing UI renders nothing), while checkout/portal
-  // throw — a write must reach the gateway.
-  async getBilling(): Promise<controlPlane.BillingSummary | null> {
-    if (!this.cp) return null;
-    return controlPlane.getBilling(this.cp);
-  }
-  async createCheckout(
-    interval: "monthly" | "annual",
-  ): Promise<controlPlane.BillingCheckout> {
-    if (!this.cp) throw new Error("Billing needs the hosted gateway.");
-    return controlPlane.createCheckout(this.cp, interval);
-  }
-  async createPortal(): Promise<controlPlane.BillingCheckout> {
-    if (!this.cp) throw new Error("Billing needs the hosted gateway.");
-    return controlPlane.createPortal(this.cp);
-  }
-
-  // ---- per-agent assignments + integration grants (multiplayer) ----
-  async setAgentAssignments(
-    agentSlugOrId: string,
-    assignments: controlPlane.AgentAssignment[] | string[],
-  ): Promise<void> {
-    if (!this.cp) throw new Error("multiplayer requires the hosted gateway");
-    return controlPlane.setAgentAssignments(
-      this.cp,
-      agentSlugOrId,
-      assignments,
-    );
-  }
-  async getAgentSettings(
-    agentSlugOrId: string,
-  ): Promise<controlPlane.AgentSettings> {
-    if (!this.cp) throw new Error("multiplayer requires the hosted gateway");
-    return controlPlane.getAgentSettings(this.cp, agentSlugOrId);
-  }
-  async setAgentSettings(
-    agentSlugOrId: string,
-    settings: {
-      allowedToolkits?: string[] | null;
-      allowedModels?: string[] | null;
-    },
-  ): Promise<void> {
-    if (!this.cp) throw new Error("multiplayer requires the hosted gateway");
-    return controlPlane.setAgentSettings(this.cp, agentSlugOrId, settings);
-  }
-  async getAgentModelChoice(
-    agentSlugOrId: string,
-  ): Promise<controlPlane.AgentModelChoiceInfo | null> {
-    if (!this.cp) return null;
-    return controlPlane.getAgentModelChoice(this.cp, agentSlugOrId);
-  }
-  async setAgentModelChoice(
-    agentSlugOrId: string,
-    choice: controlPlane.AgentModelChoice,
-  ): Promise<void> {
-    if (!this.cp) throw new Error("multiplayer requires the hosted gateway");
-    return controlPlane.setAgentModelChoice(this.cp, agentSlugOrId, choice);
-  }
-  async getOrgSettings(): Promise<controlPlane.OrgSettings> {
-    if (!this.cp) throw new Error("multiplayer requires the hosted gateway");
-    return controlPlane.getOrgSettings(this.cp);
-  }
-  async setOrgSettings(settings: {
-    allowedToolkits?: string[] | null;
-    allowedModels?: string[] | null;
-  }): Promise<void> {
-    if (!this.cp) throw new Error("multiplayer requires the hosted gateway");
-    return controlPlane.setOrgSettings(this.cp, settings);
-  }
-  async orgAudit(
-    opts: { before?: number; limit?: number } = {},
-  ): Promise<controlPlane.AuditEntry[]> {
-    if (!this.cp) throw new Error("multiplayer requires the hosted gateway");
-    return controlPlane.orgAudit(this.cp, opts);
-  }
-  async orgUsage(days: number): Promise<controlPlane.UsageRow[]> {
-    if (!this.cp) throw new Error("multiplayer requires the hosted gateway");
-    return controlPlane.orgUsage(this.cp, days);
-  }
-
-  // Grants degrade gracefully: `null` means "this deployment has no grants
-  // model" (the legacy engine path, or a host that 404s the route), which the UI
-  // treats as unsupported rather than a hard failure. A host that serves grants
-  // (the local/self-host TS host, or the cloud gateway) answers with the set.
-  async agentIntegrationGrants(
-    agentSlugOrId: string,
-  ): Promise<string[] | null> {
-    if (!this.cp) return null;
-    return controlPlane.agentIntegrationGrants(this.cp, agentSlugOrId);
-  }
-  async setAgentIntegrationGrants(
-    agentSlugOrId: string,
-    toolkits: string[],
-  ): Promise<void> {
-    // Grants are cloud-only: off-cloud there is no per-agent grants model, so
-    // this stays a no-op (never a network call) — the guard is preserved.
-    if (!this.cp) return;
-    // Delegate the WRITE to the SDK (migration wave 2a): its IntegrationsClient
-    // issues the identical `PUT /v1/agents/:id/integration-grants` with body
-    // `{toolkits}` over the SAME shared gateway fetch (bearer + `x-houston-org`)
-    // and does NOT read the response or refetch — byte-identical to the old
-    // `controlPlane.setAgentIntegrationGrants`. The READ (`agentIntegrationGrants`
-    // above) stays on cpFetch: its GET carries `transientRetryFetch`, which the
-    // SDK path lacks, so delegating it would drop that retry resilience.
-    await this.engineSdk.integrations.setGrants(agentSlugOrId, toolkits);
-  }
-
-  // ---- lifecycle no-ops the shell calls ----
-  async startAgentWatcher(): Promise<void> {}
-  async stopAgentWatcher(): Promise<void> {}
-  async startRoutineScheduler(): Promise<void> {}
-  async stopRoutineScheduler(): Promise<void> {}
-  async syncRoutineScheduler(): Promise<void> {}
-
-  wsUrl(): string {
-    return "";
-  }
-
-  /** @internal — exposed so the WS adapter can identify the default agent. */
-  defaultAgentId(): string {
-    return DEFAULT_AGENT_ID;
-  }
-}
+export class HoustonClient extends Composed {}
